Fixed booking service's method and other classes

+++ rentit/src/main/java/kg/neobis/rentit/service/BookingService.java
@@ -8,6 +8,7 @@ import kg.neobis.rentit.entity.Calendar;
 import kg.neobis.rentit.enums.BookingStatus;
 import kg.neobis.rentit.exception.AlreadyExistException;
 import kg.neobis.rentit.exception.BadRequestException;
+import kg.neobis.rentit.exception.ProductViolationException;
 import kg.neobis.rentit.exception.ResourceNotFoundException;
 import kg.neobis.rentit.repository.*;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +16,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.YearMonth;
@@ -39,7 +41,7 @@ public class BookingService {
     public List<UserBookingDto> getUserBookings() {
         User user = getAuthentication();
 
-        return user.getBookings().stream()
+        return user == null ? null : user.getBookings().stream()
                 .map(
                         entity -> {
                             UserBookingDto dto = new UserBookingDto();
@@ -47,7 +49,7 @@ public class BookingService {
                             dto.setBookingId(entity.getId());
                             dto.setStatus(entity.getBookingStatus().getStatus());
                             dto.setProductId(entity.getProduct().getId());
-                            dto.setPrice(entity.getProduct().getPrice());
+                            dto.setTotalPrice(entity.getTotalPrice());
                             dto.setProductTitle(entity.getProduct().getTitle());
                             dto.setMainImageUrl(returnMainImageUrl(entity));
 
@@ -75,7 +77,7 @@ public class BookingService {
                     dto.setBookingId(entity.getId());
                     dto.setClientId(entity.getUser().getId());
                     dto.setClientName(entity.getUser().getLastName() + " " + entity.getUser().getFirstName());
-                    dto.setPrice(entity.getProduct().getPrice());
+                    dto.setTotalPrice(entity.getTotalPrice());
                     dto.setProductTitle(entity.getProduct().getTitle());
                     dto.setProductId(entity.getProduct().getId());
                     dto.setBookDateFrom(entity.getDateFrom());
@@ -94,7 +96,7 @@ public class BookingService {
 
         if (imageProduct != null) {
             Image image = imageProduct.getImage();
-            if(image != null && image.getUrl().startsWith("http")) {
+            if (image != null && image.getUrl().startsWith("http")) {
                 return image.getUrl().replace("http", "https");
             }
         }
@@ -102,23 +104,8 @@ public class BookingService {
         return "";
     }
 
+    @Transactional
     public BookingRegistrationDto bookProduct(BookingRegistrationDto dto) {
-        LocalDate from = dto.getDateFrom();
-        LocalDate till = dto.getDateTill();
-
-        for (Booking entity : bookingRepository.findAll()) {
-            LocalDate efrom = entity.getDateFrom();
-            LocalDate etill = entity.getDateTill();
-            if(entity.getBookingStatus().equals(BookingStatus.ACCEPTED)) {
-                if ((from.compareTo(efrom) >= 0 && from.compareTo(etill) < 0) ||
-                        (from.compareTo(efrom) > 0 && from.compareTo(etill) <= 0) ||
-                        (till.compareTo(efrom) >= 0 && till.compareTo(etill) < 0) ||
-                        (till.compareTo(efrom) > 0 && till.compareTo(etill) <= 0)) {
-                    throw new AlreadyExistException("The given dates are already booked.");
-                }
-            }
-        }
-
         User user = getAuthentication();
 
         if (user == null) {
@@ -130,8 +117,35 @@ public class BookingService {
                         () -> new ResourceNotFoundException("Product was not found with ID: " + dto.getProductId())
                 );
 
-        if(user.getId().equals(product.getUser().getId())) {
-            throw new BadRequestException("You cannot add your product to favorites.");
+        LocalDate sumFrom = dto.getDateFrom();
+        LocalDate sumTill = dto.getDateTill();
+
+        int days = 1;
+
+        while (!sumFrom.isEqual(sumTill)) {
+            sumFrom = sumFrom.plusDays(1);
+            days++;
+        }
+
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(dto.getDateFrom(),
+                dto.getDateTill(), product.getId());
+
+        if (days != calendar.size()) {
+            throw new BadRequestException("Incorrect date input. Some of the given dates are not eligible" +
+                    " for booking.");
+        }
+
+        for (Calendar entity : calendar) {
+            if (entity.isBooked()) {
+                throw new AlreadyExistException("The date is already booked: " + entity.getDate());
+            } else {
+                entity.setBooked(true);
+            }
+            calendarRepository.save(entity);
+        }
+
+        if (user.getId().equals(product.getUser().getId())) {
+            throw new BadRequestException("You cannot book your product.");
         }
 
         Booking booking = new Booking();
@@ -141,8 +155,8 @@ public class BookingService {
         booking.setDateFrom(dto.getDateFrom());
         booking.setDateTill(dto.getDateTill());
         booking.setBookingStatus(BookingStatus.PENDING);
-        booking.setTotalPrice(dto.getTotalPrice());
         booking.setBookingDateTime(LocalDateTime.now());
+        booking.setTotalPrice(days * product.getPrice());
 
         bookingRepository.save(booking);
 
@@ -162,12 +176,12 @@ public class BookingService {
 
         int[] results = new int[days + 1];
 
-        for(Calendar entity: calendar) {
+        for (Calendar entity : calendar) {
             int d = entity.getDate().getDayOfMonth();
 
-            if(entity.isBooked()) {
+            if (entity.isBooked()) {
                 results[d] = 1;
-            } else if(!entity.isBooked()) {
+            } else if (!entity.isBooked()) {
                 results[d] = -1;
             }
         }
@@ -178,10 +192,10 @@ public class BookingService {
         schedule.put("free", new ArrayList<>());
         schedule.put("no booking", new ArrayList<>());
 
-        for(int i = 1; i < results.length; i++) {
-            if(results[i] == 1) {
+        for (int i = 1; i < results.length; i++) {
+            if (results[i] == 1) {
                 schedule.get("booked").add(i);
-            } else if(results[i] == -1) {
+            } else if (results[i] == -1) {
                 schedule.get("free").add(i);
             } else {
                 schedule.get("no booking").add(i);
@@ -196,6 +210,20 @@ public class BookingService {
                 .orElseThrow(
                         () -> new ResourceNotFoundException("Booking was not found with ID: " + bookingId)
                 );
+
+        User user = getAuthentication();
+
+        if (!booking.getUser().getId().equals(user.getId())) {
+            throw new ProductViolationException("Бронь не принадлежит вам.");
+        }
+
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(booking.getDateFrom(),
+                booking.getDateTill(), booking.getProduct().getId());
+
+        for (Calendar entity : calendar) {
+            entity.setBooked(false);
+            calendarRepository.save(entity);
+        }
 
         bookingRepository.deleteById(booking.getId());
 
@@ -212,15 +240,11 @@ public class BookingService {
 
         bookingRepository.save(booking);
 
-        List<Calendar> calendar = booking.getProduct().getCalendars();
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(booking.getDateFrom(),
+                booking.getDateTill(), booking.getProduct().getId());
 
-        for(Calendar entity: calendar) {
-            if((entity.getDate().isEqual(booking.getDateFrom()) ||
-            entity.getDate().isEqual(booking.getDateTill())) ||
-            entity.getDate().isAfter(booking.getDateFrom()) && entity.getDate().isBefore(booking.getDateTill())) {
-                entity.setBooked(true);
-                entity.setUser(booking.getUser());
-            }
+        for (Calendar entity : calendar) {
+            entity.setUser(booking.getUser());
             calendarRepository.save(entity);
         }
 
@@ -234,6 +258,14 @@ public class BookingService {
                 );
 
         booking.setBookingStatus(BookingStatus.REJECTED);
+
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(booking.getDateFrom(),
+                booking.getDateTill(), booking.getProduct().getId());
+
+        for (Calendar entity : calendar) {
+            entity.setBooked(false);
+            calendarRepository.save(entity);
+        }
 
         bookingRepository.save(booking);
